requirements.txt
ultralytics==8.2.103
opencv-python==4.10.0.84
mediapipe==0.10.14
easyocr==1.7.1
numpy==1.26.4
pandas==2.2.2
matplotlib==3.8.4
Pillow==10.4.0
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
pyyaml==6.0.1
streamlit==1.37.0
seaborn==0.13.2